adjust project filename && namespace, update framework v4.5.2 -> v4.8
