+++ 163MusicLyrics.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26730.16
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "163MusicLyrics", "WindowsFormsApp1\163MusicLyrics.csproj", "{A8897401-6EE2-4C86-A56A-59CCF37A5974}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "163MusicLyrics", "Application\163MusicLyrics.csproj", "{A8897401-6EE2-4C86-A56A-59CCF37A5974}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
