+++ Application/Program.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Windows.Forms;
+
+namespace Application
+{
+    static class Program
+    {
+        /// <summary>
+        /// 应用程序的主入口点。
+        /// </summary>
+        [STAThread]
+        static void Main()
+        {
+            System.Windows.Forms.Application.EnableVisualStyles();
+            System.Windows.Forms.Application.SetCompatibleTextRenderingDefault(false);
+            System.Windows.Forms.Application.Run(new MainForm());
+        }
+    }
+}
