+++ Application/Properties/AssemblyInfo.cs
