--- WindowsFormsApp1/Program.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Windows.Forms;
-
-namespace 网易云歌词提取
-{
-    static class Program
-    {
-        /// <summary>
-        /// 应用程序的主入口点。
-        /// </summary>
-        [STAThread]
-        static void Main()
-        {
-            Application.EnableVisualStyles();
-            Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new MainForm());
-        }
-    }
-}
